requirements.txt
streamlit
numpy
biopython
matplotlib
seaborn

